Reorganize file to avoid loop import
